apps/backend/requirements.txt


# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Auth & Security
Flask-JWT-Extended==4.6.0
authlib==1.2.1
python-jose[cryptography]==3.3.0

# Database
psycopg
alembic==1.12.1

# API & Serialization
marshmallow==3.20.1
flask-restx==1.3.0

# Development
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0